再帰呼び出し

+++ kotlin-tranning/task/200709.kt
@@ -1,10 +1,10 @@
-fun factorial(number: Int): Int {
-    if (number == 0) return 1
-    else  return number * factorial(number - 1) 
+fun factorial(number: Long): Long {
+    if (number == 0L) return 1L
+    else  return number * factorial(number - 1L)
 }
 
 fun main(args: Array<String>) {
-    val n = 10
+    val n = 10L
     println(factorial(n))
 }
